+++ magda-web-client/src/Components/Distribution.js
