+++ magda-web-client/src/Components/Distribution.scss
