+++ magda-web-client/src/UI/DataPreviewVega.js
@@ -113,7 +113,6 @@ class DataPreviewVega extends Component {
   }
 
   render(){
-    console.log(this.props)
     return <div className='data-preview-vega'>
             {this.renderCharts()}
             {this.renderConfigForm()}
